services/coupon/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
pydantic-settings
fastapi-pagination
gunicorn
PyJWT
httpx

